requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
pingouin>=0.5.4
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.34.0
